+++ extensions/quantumcast/src/open-cloudflow-manual.tsx
@@ -1,0 +1,6 @@
+import { open, getPreferenceValues } from '@raycast/api';
+
+export default async () => {
+  const { cloudflowBaseUrl } = getPreferenceValues();
+  await open(new URL(`${cloudflowBaseUrl}/manual/manual.html`).toString());
+};