+++ extensions/music/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Apple Music Changelog
 
+## [Update] - 2023-03-01
+
+- Fixed dislike song command
+
 ## [Updates] - 2023-02-10
 
 - Fixed typo on "add to playlist"